+++ Lib/fontbakery/profiles/fontbureau.py
@@ -9,6 +9,7 @@ PROFILE = {
         "hinting_impact",
         "inconsistencies_between_fvar_stat",
         "no_mac_entries",
+        "render_own_name",
         "vtt_volt_data",  # very similar to vttclean, may be a good idea to merge them.
         "vttclean",
     ],