+++ Lib/fontbakery/profiles/microsoft.py
@@ -22,6 +22,7 @@ PROFILE = {
         "inconsistencies_between_fvar_stat",
         "no_debugging_tables",
         "no_mac_entries",
+        "render_own_name",
         "typoascender_exceeds_Agrave",
         "ufolint",
         "ufo_features_default_languagesystem",